+++ Tests/AwsLambdaFunction.Sqs.GroupB/AwsLambdaFunction.Sqs.GroupB.Api/ClientCreatedEventConsumer.cs
@@ -13,7 +13,7 @@ using Microsoft.Extensions.Logging;
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: IntentTemplate("Intent.Aws.Lambda.Functions.Sqs.LambdaFunctionConsumer", Version = "1.0")]
 
-namespace Aws.Sqs.GrpB.Api
+namespace Lambda
 {
     public class ClientCreatedEventConsumer
     {