+++ Tests/AwsLambdaFunction.Sqs.GroupB/AwsLambdaFunction.Sqs.GroupB.Api/Configuration/ApplicationSecurityConfiguration.cs
@@ -1,4 +1,4 @@
-using Aws.Sqs.GrpB.Api.Services;
+using AwsLambdaFunction.Sqs.GroupB.Api.Services;
 using AwsLambdaFunction.Sqs.GroupB.Application.Common.Interfaces;
 using Intent.RoslynWeaver.Attributes;
 using Microsoft.Extensions.Configuration;
@@ -7,7 +7,7 @@ using Microsoft.Extensions.DependencyInjection;
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: IntentTemplate("Intent.Application.Identity.ApplicationSecurityConfiguration", Version = "1.0")]
 
-namespace Aws.Sqs.GrpB.Api.Configuration
+namespace AwsLambdaFunction.Sqs.GroupB.Api.Configuration
 {
     public static class ApplicationSecurityConfiguration
     {