+++ Tests/AwsLambdaFunction/AwsLambdaFunction.Api/DynClientsFunctions.cs
@@ -21,7 +21,7 @@ using Microsoft.Extensions.Logging;
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: IntentTemplate("Intent.Aws.Lambda.Functions.LambdaFunctionClassTemplate", Version = "1.0")]
 
-namespace AwsLambdaFunction.Api
+namespace Lambda
 {
     public class DynClientsFunctions
     {